Create Manhattan plot from rTASSEL association output

This function allows for quick generation of a Manhattan plot from rTASSEL association statistical output data.

manhattanPlot(
  assocStats,
  trait,
  threshold,
  colors = c("#91baff", "#3e619b"),
  showSigMarkers = FALSE,
  showRug = FALSE
)

Arguments

assocStats

rTASSEL association statistical output. This output is derived from association output dat frames ending in _Stats. This data contains the necessary column variables for this plot to work.

trait

Which phenotypic trait do you want to plot?

threshold

User-defined \(-log_{10}(p)\)-value threshold for significant marker determination. Once specified any marker points higher than this line will be highlighted.

colors

A vector of colors used for differentiating multiple chromosomes. Defaults to 2 shades of blue.

showSigMarkers

Should significantly highlighted markers be labelled? If TRUE, marker points will labelled with their respective marker labels. Defaults to FALSE.

showRug

Should a rug plot be plotted under the Manhattan plot? Defaults to FALSE.

Value

Returns a ggplot2 object