readPhenotypeFromDataFrame {rTASSEL}	R Documentation

Wrapper function of TasselGenotypePhenotype class for phenotype data from an R data frame

Description

This function is a wrapper for the TasselGenotypePhenotype class. It is used for storing phenotype information into a class object.

Usage

readPhenotypeFromDataFrame(phenotypeDF, taxaID, attributeTypes = NULL)

Arguments

phenotypeDF
A R object of class data.frame.

taxaID
The column name that represents your taxa data as a string.

attributeTypes
A vector of non-taxa attributes. If NULL, all attributes will be TASSEL <data> types.

Value

Returns an object of TasselGenotypePhenotype class.

[Package rTASSEL version 0.9.32 Index]