startLogger {rTASSEL}	R Documentation

Start TASSEL logging information

Description

This function will create a file for storing logging output from TASSEL.

Usage

startLogger(fullPath = NULL, fileName = NULL, ...)

Arguments

fullPath
full working path of log file location. If NULL, logging file will be added to current working directory.

fileName
name of logging file. If NULL, filename will resort to "rTASSEL_log".

...
additional parameters to be added

[Package rTASSEL version 0.9.32 Index]